requirement.txt
fastapi==0.116.1
pandas==2.3.2
tensorflow==2.20.0
uvicorn==0.35.0
numpy==1.26.0
Pillow==10.0.0
